fix: 🐛 extract getting application status
move logic to own class that is not dependant on MyPages class. hopeing
to reduced response time
