+++ app/libs/__init__.py
@@ -1,4 +1,10 @@
-from .classes import Session, DataClassApplication, Viva, VivaMyPages, VivaApplication, VivaAttachments
+from .classes import Session
+from .classes import DataClassApplication
+from .classes import Viva
+from .classes import VivaMyPages
+from .classes import VivaApplication
+from .classes import VivaApplicationStatus
+from .classes import VivaAttachments
 
 from .personal_number_helper import hashids_instace, hash_to_personal_number
 from .datetime_helper import milliseconds_to_date_string
