+++ app/libs/classes/__init__.py
@@ -4,3 +4,4 @@ from .viva import Viva
 from .viva_my_pages import VivaMyPages
 from .viva_attachments import VivaAttachments
 from .viva_application import VivaApplication
+from .viva_application_status import VivaApplicationStatus
