+++ app/libs/classes/viva_application_status.py
@@ -1,0 +1,101 @@
+from zeep.exceptions import Fault
+
+from . import Viva
+
+
+class VivaApplicationStatus(Viva):
+    def __init__(self, wsdl='VivaApplication', personal_number=None):
+        super(VivaApplicationStatus, self).__init__()
+
+        self._service = self._get_service(wsdl)
+
+        if not isinstance(personal_number, str) or len(personal_number) < 11:
+            raise Fault(message='User missing', code=400)
+
+        self._personal_number = personal_number
+
+    def get(self):
+        """
+        ApplicationStatus förklaring:
+
+        -1 - fel (t.ex. person finns inte i personregistret)
+
+        eller summan av:
+        1 - ansökan tillåten
+        2 - autosave finns (man har påbörjat en ansökan)
+        4 - väntar signering (avser medsökande: sökande har signerat en ansökan som inkluderar medsökande)
+        8 - väntar att medsökande ska signera (avser sökande medan hen väntar på att medsökande ska signera)
+        16 - ansökan inskickad
+        32 - ansökan mottagen(/registrerad i Viva)
+        64 - komplettering begärd
+        128 - ärende finns (försörjningsstödsärende)
+        256 - ett ärende är aktiverat på web (egenskap på ärendet som gör att det visas på Mina sidor)
+        512 - ärendet tillåter e-ansökan  (egenskap på ärendet som gör att det går att skapa fortsatt ansökan)
+
+        2, 4 och 8 blir bara aktuella när man parkerar eller autosparar ansökan i Viva.
+
+        Det är bara när ApplicationStatus innehåller 1 (summan är udda) som ansökan kan lämnas in.
+
+        128 + 256 + 512 kommer i princip alltid att vara med eftersom ett ärende som tillåter e-ansökan är en förutsättning för fortsatt ansökan.
+
+        När ApplicationStatus bara är 1 finns inget ärende och man kan lämna in en grundansökan.
+        """
+
+        status_number = self._service.APPLICATIONSTATUS(
+            SUSER=self._personal_number,
+            SPNR=self._personal_number,
+            SCASETYPE='01',  # 01 = EKB
+            SSYSTEM=1
+        )
+
+        status_description = {
+            1: 'Ansökan tillåten',
+            2: 'Autosave',
+            4: 'Väntar signering',
+            8: 'Väntar att medsökande ska signera',
+            16: 'Ansökan inskickad',
+            32: 'Ansökan mottagen/registrerad',
+            64: 'Komplettering begärd',
+            128: 'Arende finns (försörjningsstödsärende)',
+            256: 'Ett ärende är aktiverat på webben',
+            512: 'Ärendet tillåter e-ansökan',
+        }
+
+        statuses = list()
+
+        if status_number > 128:
+            key = 128
+            statuses.append({
+                'code': key,
+                'description': status_description[key]
+            })
+            status_number -= key
+
+        if status_number > 256:
+            key = 256
+            statuses.append({
+                'code': key,
+                'description': status_description[key]
+            })
+            status_number -= key
+
+        if status_number > 512:
+            key = 512
+            statuses.append({
+                'code': key,
+                'description': status_description[key]
+            })
+            status_number -= key
+
+        if status_number in status_description:
+            statuses.append({
+                'code': status_number,
+                'description': status_description[status_number]
+            })
+        else:
+            statuses.append({
+                'code': status_number,
+                'description': 'N/A'
+            })
+
+        return self._helpers.serialize_object(statuses)
